+++ zerver/views/user_groups.py
@@ -18,11 +18,7 @@ from zerver.actions.user_groups import (
     do_update_user_group_name,
     remove_subgroups_from_user_group,
 )
-from zerver.decorator import (
-    require_human_non_guest_user,
-    require_non_guest_user,
-    require_user_group_create_permission,
-)
+from zerver.decorator import require_non_guest_user, require_user_group_create_permission
 from zerver.lib.exceptions import JsonableError
 from zerver.lib.mention import MentionBackend, silent_mention_syntax_for_user
 from zerver.lib.response import json_success
@@ -131,7 +127,7 @@ def get_user_groups(
 
 
 @transaction.atomic(durable=True)
-@require_human_non_guest_user
+@require_non_guest_user
 @typed_endpoint
 def edit_user_group(
     request: HttpRequest,
@@ -229,7 +225,7 @@ def deactivate_user_group(
     return json_success(request)
 
 
-@require_human_non_guest_user
+@require_non_guest_user
 @typed_endpoint
 @transaction.atomic(durable=True)
 def update_user_group_backend(
@@ -506,7 +502,7 @@ def remove_subgroups_from_group_backend(
     return json_success(request)
 
 
-@require_human_non_guest_user
+@require_non_guest_user
 @typed_endpoint
 @transaction.atomic(durable=True)
 def update_subgroups_of_user_group(